Session usage is implemented.

+++ src/main/java/com/codecool/microservices/model/User.java
@@ -2,14 +2,14 @@ package com.codecool.microservices.model;
 
 public class User {
 
-    private long id;
+    private int id;
     private String firstName;
     private String lastName;
     private String email;
     private String address;
     private String phoneNumber;
 
-    public User(long id, String firstName, String lastName, String email, String address, String phoneNumber) {
+    public User(int id, String firstName, String lastName, String email, String address, String phoneNumber) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
